+++ CHANGELOG.md
@@ -20,6 +20,8 @@ This document follows the conventions laid out in [Keep a CHANGELOG][].
 -   Fixed crash on exit of the Python interpreter if a python class
     derived from a .NET class has a `__namespace__` or `__assembly__`
     attribute (#481)
+-   Fixed conversion of 'float' and 'double' values (#486)
+
 
 ## [2.3.0][] - 2017-03-11